Add support for HTTP-only filters and modifying status code
The new IHttpFunctionInvocationFilter is only called for HTTP-triggered functions and allows inspecting the currently running function with much more data than the provided IFunctionInvocationFilter provides (i.e. custom attributes on the method, as well as its descriptor), in addition to accessing the HTTP context itself.

The HttpStatusCodeException (and its accompanying filter) in turn, allow those HTTP filters to throw that specific exception, which causes a subsequent IFunctionExceptionFilter to run and modify the response accordingly.

Finally, a minor refactoring on the service conventions allows us to move all that convention code to an extension method (for now).